Moves overview related files to their own folder
